requirements.txt
langchain
python-dotenv
langchain_community
ipykernel
pypdf
arxiv
langchain-text-splitters
langchain-openai

chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
chromadb
langchain_chroma

crewai
crewai_tools
beautifulsoup4
langchain_groq
langchain_core
langserve